LPS-61952 Move com.liferay.portlet.journal.util to com.liferay.journal.kernel.util
